# backend/docs/swagger.yaml
definitions:
...
  github_com_raids-lab_crater_internal_resputil.Response-internal_handler_UserBanPolicyReq:
    properties:
      code:
        description: 依然保持 int (ErrorCode) 类型
        type: integer
      data:
        $ref: '#/definitions/internal_handler.UserBanPolicyReq'
      msg:
        type: string
    type: object
...
  internal_handler.UserBanPolicyReq:
    properties:
      allowDatasetDownload:
        type: boolean
      allowImageBuild:
        type: boolean
      allowJobSubmission:
        type: boolean
      allowModelDownload:
        type: boolean
      allowPlatformAccess:
        type: boolean
    type: object
...
info:
  contact: {}
  description: This is the API server for Crater, a Multi-tenant AI Model Training
    Platform based on Kubernetes.
  title: Crater API
  version: 1.0.0
paths:
...
  /v1/admin/system-config/user-ban:
    get:
      description: 获取封禁用户仍被允许使用的平台能力
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-internal_handler_UserBanPolicyReq'
      security:
      - Bearer: []
      summary: 获取用户封禁能力策略
      tags:
      - SystemConfig
    put:
      consumes:
      - application/json
      description: 设置封禁用户是否仍可访问平台、提交作业、制作镜像以及下载模型或数据集
      parameters:
      - description: 封禁能力策略
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UserBanPolicyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-internal_handler_UserBanPolicyReq'
      security:
      - Bearer: []
      summary: 更新用户封禁能力策略
      tags:
      - SystemConfig
...
securityDefinitions:
  Bearer:
    description: 访问 /login 并获取 TOKEN 后，填入 'Bearer ${TOKEN}' 以访问受保护的接口
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"